llm/usecase/finetune/requirements.txt
# server requirements
fedops
scikit-learn
transformers
trl
peft
bitsandbytes
hydra-core
numpy==1.26.4